imxrt10xx: Use port state to handle interrupts
A GPIO port already knows what kind of port it is. Passing in the
GPIO port value to handle_interrupt is redundant. The commit removes
the Port::handle_interrupt argument, and discerns the GPIO port
using a check for pointer equality.

This would be a bit easier if we could equate StaticRefs. The next
commit will propose that change. We can equate (and order) pointers
in Rust, so it should take a simple derive(PartialEq, Eq) on the
wrapper type to make it work.