kernel,imxrt10xx: Equate StaticRefs
The previous commit introduced a match statement on pointer equality.
It would be simpler to express the operation by equating StaticRefs.
We can fully equate pointers in Rust, so we just derive the traits
on StaticRef.

See the previous commit for the manual implementation of this match
statement. If we don't want to extend StaticRef, I'll revert this
commit.