imxrt10xx: configure UART clock in peripheral init
We need to disable _all_ UART clocks at the clock gates before we
change the UART clock root. Before this commit, the configure method
would disable the clock gate for the current instance, configure the
clocks for _all_ UART instances, then re-enable its own clock. Though
the setting was the same for each configure call, so there were
probably no ill effects, we shouldn't configure the UART clock root in
each instance's configure call. This commit moves the UART root clock
configuration into each board's early setup routine. This should be
equivalent to what we were doing before.

The 'disable' term for the clock mux is misleading. This call actually
transitions the UART to use PLL3, not disable the clock to the UART.
(If the PLL is off, that's not obvious from this call.) The commit
refactors the CCM API to expose setters and getters.