imxrt10xx: refactor GPT for board instantiation
When we introduced the teensy40 board, we extended the GPT timers
to support different input clocks. But, without board-based periph
instantiation, there was not a clean way to statically specify the
clock frequencies, which were changing board-by-board.

Now that we have board-based peripheral instantiation, we can easily
specify the GPT clock frequency statically, for each board. The GPT
timer is generic over a frequency. You're expected to supply a valid
Frequency, and configure your timer to run at that frequency in the
board setup.